move osmscout/gpx includes to osmscoutgpx
