+++ Demos/include/PositionSimulator.h
@@ -20,7 +20,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include <osmscout/gpx/GpxFile.h>
+#include <osmscoutgpx/GpxFile.h>
 
 #include <QObject>
 #include <QString>
