+++ Demos/src/ElevationProfile.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include <osmscout/gpx/Import.h>
+#include <osmscoutgpx/Import.h>
 
 #include <osmscout/util/CmdLineParsing.h>
 #include <osmscout/util/Logger.h>
@@ -26,7 +26,7 @@
 
 #include <iostream>
 #include <osmscout/Database.h>
-#include <osmscout/gpx/Export.h>
+#include <osmscoutgpx/Export.h>
 
 struct Arguments
 {
