+++ Demos/src/GpxPipe.cpp
@@ -17,8 +17,8 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include <osmscout/gpx/Import.h>
-#include <osmscout/gpx/Export.h>
+#include <osmscoutgpx/Import.h>
+#include <osmscoutgpx/Export.h>
 
 #include <osmscout/util/CmdLineParsing.h>
 #include <osmscout/util/Logger.h>
