+++ Demos/src/PositionSimulator.cpp
@@ -21,7 +21,7 @@
 
 #include <osmscout/util/Logger.h>
 
-#include <osmscout/gpx/Import.h>
+#include <osmscoutgpx/Import.h>
 
 static const std::chrono::milliseconds TickDuration(100);
