+++ libosmscout-gpx/include/osmscoutgpx/Extensions.h
@@ -1,9 +1,9 @@
-#ifndef OSMSCOUT_GPX_ROUTE_H
-#define OSMSCOUT_GPX_ROUTE_H
+#ifndef OSMSCOUT_GPX_EXTENSIONS_H
+#define OSMSCOUT_GPX_EXTENSIONS_H
 
 /*
   This source is part of the libosmscout-gpx library
-  Copyright (C) 2017 Lukas Karas
+  Copyright (C) 2021 Lukas Karas
 
   This library is free software; you can redistribute it and/or
   modify it under the terms of the GNU Lesser General Public
@@ -20,23 +20,36 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/gpx/TrackPoint.h>
+#include <osmscoutgpx/GPXImportExport.h>
 
-#include <osmscout/gpx/GPXImportExport.h>
-
-#include <vector>
 #include <string>
-#include <optional>
+#include <vector>
 
 namespace osmscout {
 namespace gpx {
 
-class OSMSCOUT_GPX_API Route {
+class OSMSCOUT_GPX_API Extensions {
+
 public:
-  std::optional<std::string> name;
-  std::vector<TrackPoint> points;
+  Extensions() = default;
+
+  class Element {
+  public:
+    explicit Element(const std::string& ns, const std::string& elemName, const std::string& elemValue);
+
+    const std::string& GetName() const { return name; }
+    const std::string& GetValue() const { return value; }
+
+  private:
+    std::string name;
+    std::string value;
+
+  };
+
+  std::vector<Element> elements;
+
 };
 }
 }
 
-#endif //OSMSCOUT_GPX_ROUTE_H
+#endif //OSMSCOUT_GPX_EXTENSIONS_H
