+++ libosmscout-gpx/include/osmscoutgpx/GPXFeatures.h.cmake
