+++ libosmscout-gpx/include/osmscoutgpx/TrackPoint.h
