+++ libosmscout-gpx/include/osmscoutgpx/Utils.h
@@ -19,8 +19,8 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/gpx/GPXImportExport.h>
-#include <osmscout/gpx/TrackPoint.h>
+#include <osmscoutgpx/GPXImportExport.h>
+#include <osmscoutgpx/TrackPoint.h>
 
 #include <string>
 #include <memory>
