+++ libosmscout-gpx/include/osmscoutgpx/Waypoint.h
@@ -1,9 +1,9 @@
-#ifndef OSMSCOUT_GPX_EXTENSIONS_H
-#define OSMSCOUT_GPX_EXTENSIONS_H
+#ifndef OSMSCOUT_GPX_WAYPOINT_H
+#define OSMSCOUT_GPX_WAYPOINT_H
 
 /*
   This source is part of the libosmscout-gpx library
-  Copyright (C) 2021 Lukas Karas
+  Copyright (C) 2017 Lukas Karas
 
   This library is free software; you can redistribute it and/or
   modify it under the terms of the GNU Lesser General Public
@@ -20,36 +20,37 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/gpx/GPXImportExport.h>
+#include <osmscoutgpx/TrackPoint.h>
+
+#include <osmscout/GeoCoord.h>
+
+#include "GPXImportExport.h"
 
 #include <string>
-#include <vector>
+#include <optional>
 
 namespace osmscout {
 namespace gpx {
 
-class OSMSCOUT_GPX_API Extensions {
-
+class OSMSCOUT_GPX_API Waypoint {
 public:
-  Extensions() = default;
+  explicit Waypoint(GeoCoord coord) :
+      coord(coord) {
+  }
 
-  class Element {
-  public:
-    explicit Element(const std::string& ns, const std::string& elemName, const std::string& elemValue);
+  std::optional<std::string> name;
+  std::optional<std::string> description;
+  std::optional<std::string> symbol;
 
-    const std::string& GetName() const { return name; }
-    const std::string& GetValue() const { return value; }
-
-  private:
-    std::string name;
-    std::string value;
-
-  };
-
-  std::vector<Element> elements;
-
+  osmscout::GeoCoord coord;
+  std::optional<double> elevation; // meters above sea
+  std::optional<Timestamp> time;
+  std::optional<double> course; // alias magvar - degrees, 0.0 <= value < 360.0
+  std::optional<double> hdop; // meters
+  std::optional<double> vdop; // meters
+  std::optional<double> pdop; // meters
 };
 }
 }
 
-#endif //OSMSCOUT_GPX_EXTENSIONS_H
+#endif //OSMSCOUT_GPX_WAYPOINT_H
