+++ libosmscout-gpx/include/osmscoutgpx/meson.build
