+++ libosmscout-gpx/include/osmscoutgpx/private/meson.build
