+++ libosmscout-gpx/src/meson.build
@@ -1,15 +1,15 @@
 osmscoutgpxSrc = [
-            'src/osmscout/gpx/TrackSegment.cpp',
-            'src/osmscout/gpx/GpxFile.cpp',
-            'src/osmscout/gpx/Utils.cpp',
-            'src/osmscout/gpx/Track.cpp',
-            'src/osmscout/gpx/Extensions.cpp',
+            'src/osmscoutgpx/TrackSegment.cpp',
+            'src/osmscoutgpx/GpxFile.cpp',
+            'src/osmscoutgpx/Utils.cpp',
+            'src/osmscoutgpx/Track.cpp',
+            'src/osmscoutgpx/Extensions.cpp',
           ]
 
 if xml2Dep.found()
   osmscoutgpxSrc += [
-            'src/osmscout/gpx/Import.cpp',
-            'src/osmscout/gpx/Export.cpp',
+            'src/osmscoutgpx/Import.cpp',
+            'src/osmscoutgpx/Export.cpp',
           ]
 endif
