+++ libosmscout-gpx/src/osmscoutgpx/Extensions.cpp
@@ -1,6 +1,6 @@
 /*
   This source is part of the libosmscout-gpx library
-  Copyright (C) 2017 Lukas Karas
+  Copyright (C) 2021 Lukas Karas
 
   This library is free software; you can redistribute it and/or
   modify it under the terms of the GNU Lesser General Public
@@ -17,4 +17,17 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/gpx/GpxFile.h>
+#include <osmscoutgpx/Extensions.h>
+
+#include <osmscout/system/Assert.h>
+
+using namespace osmscout::gpx;
+
+Extensions::Element::Element(const std::string &ns, const std::string &elemName, const std::string &elemValue)
+  : name(), value(elemValue)
+{
+  size_t sz = ns.size() + 1 + elemName.size();
+  assert(sz > 1);
+  name.reserve(sz);
+  name.append(ns).append(":").append(elemName);
+}
