+++ libosmscout-gpx/src/osmscoutgpx/Utils.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/gpx/Utils.h>
+#include <osmscoutgpx/Utils.h>
 
 #include <osmscout/util/Geometry.h>
 #include <osmscout/util/Logger.h>
